+++ I18n/backOffice/default/en_US.php
@@ -1,0 +1,10 @@
+<?php
+
+return array(
+    'Help !!!' => 'Help !!!',
+    'If you are logged in with another customer he will be offline' => 'If you are logged in with another customer he will be offline',
+    'Take customer account' => 'Take customer account',
+    'To ask for help' => 'To ask for help',
+    'To connect to the customer\'s account' => 'To connect to the customer\'s account',
+    'You have not need the customer password' => 'You have not need the customer password',
+);