+++ CHANGELOG.md
@@ -1,3 +1,17 @@
 # 1.0.1
 
 - Remove old cart in the session
+
+# 1.0.2
+
+- #2 Fix checkAuth in controller
+
+# 1.1.0
+
+- Add event `TakeCustomerAccountEvents::TAKE_CUSTOMER_ACCOUNT`
+- Add hook `take-customer-account.form`
+- Add the ability to change the redirection from event listener. For this, to throw an exception of type `RedirectException`
+- Add screenshot and logo
+- Fix error message in form
+- Fix checkAuth in template
+